fabs added, requirements are splited up.
- Added `inspects.has_method`(cls, methodName). Check if class cls has method with name methodName directly,
or in one of it's super-classes.

- Added `ymlparsers` module - `load`/`safe_dump` a Hierarchical Yaml files. This is essentially wrapper arround HiYaPyCo project with streamlined
and extended API and couple of work-arrounds.

Note: this module doesn't use any package-level variables in hiYaPyCo module.

It also has another defaults for `load`/`safe_dump` methods.
They can be overridden in `initConfig()` function.

It is mandatory to call `initConfig()` function before any method in `ymlparsers` module.